Update frontend nginx image
Changing the base image for the frontend webserver to unprivileged
version of nginx to prevent issues in environments where running
images with the root user is not possible. This change of the base
image also means that the web server will be exposed on port 8080
instead of the previous port 80 (because non-root users are not
able to allocate ports below 1024.

Adding a custom configuration for nginx to specify the mapping of
URIs to files. This is needed because with react-router all URIs
need to be mapped to the index.html file to accomplish client-side
routing via JavaScript. The added configuration is just the default
nginx configuration extended with the "try_files" directive. In the
future it would be nice to also have some config for caching and
gzip compression.