Add nginx caching and compression
Adding explicit caching for static assets. Because the file names
will have the hash of the contents in it, static files can be
cached for a long time. The index.html should never be cached so
that on newer versions of the application the browser knows that
the file names of the static assets have changed.

Also adding gzip compression to nginx to boost the performance for
larger files.